Ran first three datasets.
First three datasets ran successfully.
Results are present in the .shelf files.
I expect to re-run the analysis again, but
for both cpu and gpu modes. First, I want
to make sure I've collected the neccesary information
to build the plots for the final paper in the shelf
files. I'll start building the plots next.