Progress on neuralnet comparison plot.
Able to draw the neuralnet comparison plot
with a nice colormap, error bars, and significant
difference labels.

The significance labels are still dummies, and I need to record
the number of training cycles, individually recorded
accuracy measurements (for min and maxes) and implement
Tukey's HSD for each dataset before it will be done.